[NUI] Make AccessibilityManager APIs as public ones
- Move AccessibilityManager class from internal to public
- Added doxygen and make AccessibilityManager work in NUI
